requirements.txt
langgraph>=0.0.15
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.2
pydantic>=2.0.0
typing-extensions>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
Faker>=19.0.0
pyspark>=3.5.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
python-dateutil>=2.8.2
pytz>=2020.1
pytest>=7.0.0 